Reduce duplication of code in GH Actions And fix the Release creation steps
