+++ .github/workflows/deployment.yml
@@ -12,10 +12,10 @@ on:
         description: Version to be published
         default: v1.11.0
         required: false
-      create_release:
-        description: Create the release even if publishing already happened
-        default: 'false'
-        required: false
+      # create_release:
+      #   description: Create the release even if publishing already happened
+      #   default: 'false'
+      #   required: false
 
 jobs:
   check_version:
@@ -144,14 +144,14 @@ jobs:
 
       # Each individual action should git-add the desired files
       - name: Update Project Files with new Version
-        if: steps.check_project.outputs.files_staged == 1 || github.event.input.create_release == 'true'
+        if: steps.check_project.outputs.files_staged == 1
         run: |
           git status
           git commit -m "Version Bumped to ${{ fromJson(steps.get_version.outputs.top_package).version }}" --signoff
           git push --force-with-lease origin_token master
 
       - name: Create the Release
-        if: steps.check_project.outputs.files_staged == 1 || github.event.input.create_release == 'true'
+        if: steps.check_project.outputs.files_staged == 1
         id: create_release
         uses: actions/create-release@v1
         env:
@@ -169,8 +169,9 @@ jobs:
 
       - name: Upload the Zowe Explorer Assets
         if: |
-          (steps.check_project.outputs.files_staged == 1 || github.event.input.create_release == 'true') &&
+          steps.check_project.outputs.files_staged == 1 &&
           steps.deploy_zowe_explorer.outputs.archive != null
+        # steps.check_project.outputs.files_staged == 1 || github.event.input.create_release == 'true') &&
         uses: actions/upload-release-asset@v1
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
@@ -182,8 +183,9 @@ jobs:
 
       - name: Upload the FTP Extension for Zowe Explorer Assets
         if: |
-          (steps.check_project.outputs.files_staged == 1 || github.event.input.create_release == 'true') &&
+          steps.check_project.outputs.files_staged == 1 &&
           steps.deploy_zowe_explorer_ftp_extension.outputs.archive != null
+        # steps.check_project.outputs.files_staged == 1 || github.event.input.create_release == 'true') &&
         uses: actions/upload-release-asset@v1
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
@@ -195,8 +197,9 @@ jobs:
 
       - name: Upload the Zowe Explorer APIs Assets
         if: |
-          (steps.check_project.outputs.files_staged == 1 || github.event.input.create_release == 'true') &&
+          steps.check_project.outputs.files_staged == 1 &&
           steps.deploy_zowe_explorer_api.outputs.archive != null
+        # steps.check_project.outputs.files_staged == 1 || github.event.input.create_release == 'true') &&
         uses: actions/upload-release-asset@v1
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
@@ -208,8 +211,9 @@ jobs:
 
       - name: Upload the ESLint plugin for Zowe Explorer Assets
         if: |
-          (steps.check_project.outputs.files_staged == 1 || github.event.input.create_release == 'true') &&
+          steps.check_project.outputs.files_staged == 1 &&
           steps.deploy_eslint_plugin_zowe_explorer.outputs.archive != null
+        # steps.check_project.outputs.files_staged == 1 || github.event.input.create_release == 'true') &&
         uses: actions/upload-release-asset@v1
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
